Removing some unnecessary initialization and code paths.
Very slight change in dropped frames. NB: discovered that the actual view
being displayed affects the throughput a lot. In this case, I used a view of
a featureless surface and got 17.5% dropped frames (was 18.6% before
commit). However if drone is viewing some more objects (computer, chair and
me sitting in chair, etc) the amount of dropped frames is around 36%.