Trying to fix connect failing some of the time.
Sometimes we fail to connect to the drone. Seems like the drone actually
drops some datagrams we send. This is of course very bad when we try to
synchronize and only send one datagram that the drone must receive.

In this commit, the NavDataReader thread will send datagrams until it gets a
reply from the drone. There is also an insertion of an AT command that tells
the drone to switch to the navdata_demo mode if we receive a watchdog
time-out when in the BOOTSTRAP state. The time-out is only 50 ms so it is
quite possible that this happens every now and then when we try to connect.